api/requirements.txt
django
djangorestframework
django-cors-headers
python-chess